katzenmint: move joinNetwork to register command
